container: Make single image removal idempotent
Due to bugs or logic errors, it's possible that we might attempt
to prune an image that's already been pruned.  In order to
make things more robust, let's avoid making this a hard error.

Keep the semantics for the existing "prune multiple images" API,
but add a new one that prunes a singular image, and does not
error in the not-found case.  This can be used by higher level
software.