tar/export: Always generate `/var/tmp`
This is a bit of a hack, but basically the ostree model of
empty `/var` defers creation of basic directory structure to
e.g. `systemd-tmpfiles`, but that isn't run in containers.

For now, let's keep this as the only special case.  Perhaps down
the line we may actually need to effectively parse/execute
systemd-tmpfiles inside the stream in the future, but hopefully
we can avoid that.

Many things are going to use `/var/tmp` during image builds,
but that's much less likely with e.g. `/var/log`.

Closes: https://github.com/ostreedev/ostree-rs-ext/issues/417