Cleanly display/error if origin is rpm-ostree owned
If one does any machine-local layering/overrides via rpm-ostree,
we really should error out if one does `bootc upgrade`; otherwise
we'll just lose their changes.

This is a bit of an ugly hack; what would be more elegant is
to add something like a clear `owner` flag into origin files, and
have rpm-ostree flip that on if it needs to take over.

But for now, let's do this to avoid more confusion.

Signed-off-by: Colin Walters <walters@verbum.org>