Add a SELinux module with helper to verify install_t
This is a hack to help detect the situation in
https://github.com/containers/bootc/issues/24

Ultimately, this whole issue makes it extremely hard
to expose a *library* interface to our users because the requirement
is infectious - they also need to be `install_t`.

Anyways for now, this new module will help at least detect
the situation.