Add an `internal-tests run-privileged-integration`
Add a feature to enable internal tests.  I originally started
by matching
https://github.com/ostreedev/ostree-rs-ext/commit/6543c855580da3a3f2a85bc8b4a294b4c92c2829
but since here the binary is really end user facing and not just
a demo, we shouldn't enable the feature by default.

If in the future we split off the tiny CLI bit out of the workspace,
then we could auto-enable but that seems not worth it.

Signed-off-by: Colin Walters <walters@verbum.org>