xtask: Add timestamp to version
This is a copy of
https://github.com/containers/bootc/commit/0c8c6834f0aecdf18d8b9eec07a3810743ad0d37

I was hitting an issue with our COPR builds where we were getting
*earlier* builds because the git hash was the first component
of the version, and the leading numbers there don't increment,
and RPM wants higher versions.

Force newer-is-better by injecting the date (up to the minute, not
second because we're not going to build more than one a second).