efi: Skip adopting for systemd-boot and systemd-stub
If the system is booted with systemd-boot or systemd-stub then `bootctl`
will handle updating the boot components. Both of these components write
identifying information to EFI variables as NUL terminated UTF-16
strings. Check these variables to detect if either systemd-boot or
systemd-stub are in use and skip adoption if so.