install: Drop host bind mount code that never worked
It only appeared to kind of work because we ended up with a copy
of the host rootfs for some reason.

As the comment in the code indicates, what we're really trying
to do here is the problem domain covered by
https://brauner.io/2023/02/28/mounting-into-mount-namespaces.html
where we want to dynamically insert mounts into a running container.

I was thinking about this and with some experimentation I realized
that while we can't change the mount namespacing to make `/tmp`
in our container be the same thing as the host `/tmp`, the "magic link"
for `/proc/1/root` *will* traverse outside our mount namespace, so
it works to just symlink `/tmp` -> `/proc/1/root/tmp`.  It's not
beautiful, but it works.

Signed-off-by: Colin Walters <walters@verbum.org>