chunking: Minor comment cleanup/clarifications
The default Rust style includes a space between the comment `//`
and the text, so do that consistently.

Ensure that "documentation comments" are consistently associated
with their function.

Expand acronyms like "hs" and "ms" etc. in many places because
I find this easier to read.  (Particularly I have a strong association
with "ms" as "millisecond")