container: Have ManifestDiff borrow, make fields pub
While efficiency here doesn't matter a whole lot, I think it
is more elegant to borrow from the input arguments.

While we're here, make those fields `pub`.  This makes it easier
for callers to perform their own computations on the values instead
of just being able to invoke `print`.

Motivated by potentially using this in rpm-ostree for chunking
size analysis.

Finally, also pad the output in `print` to line up the fields
as it looks prettier.