Add previous container state manifest
This commit is a prerequisite to issue #4176 in the rpm-ostree repo,
implementing the `rpm-ostree upgrade --check` feature for ostree native
containers.

A previous_state `LayeredImageState` object is added to the
`PreparedImport` object, used to store the previous container state when
importing a container image.

A `export_as_string()` function is added to the `ManifestDiff` object to
allow rpm-ostree to print the manifest diff on the client side. The code
is exactly the same as the existing `print()` function for
`ManifestDiff`, except it returns a string instead of printing out the
diff in the terminal.