container: Move more metadata into `ExportOpts`
This drops two extra arguments we added over time; in
a few places before we had e.g. `None, None, None` being passed
which just looks awkward.  And we also threaded through all 3
in various places.

The `ExportOpts` just needs to grow a lifetime argument, but that
turned out to not be too bad when I realized we could use the
elided lifetime `<'_>` in all methods that use it.