Import static grub configuration
Pull this from coreos-assembler; prep for moving ownership
of it into bootupd.  The big change though is the addition of
sourcing a `platform.cfg` which is where we'll put the stuff
that currently lives in `platforms.yaml` in FCOS as well as
the Ignition glue.

This glue can be used as is by Sagano, where the base images don't
need the Ignition glue or platforms.yaml.