Add RequiredHostSpec
We want to gracefully handle the case where there's no booted/staged
image - i.e. when there's no spec.

However we only support deploying a container image.  So add
a `RequiredHostSpec` that is like `HostSpec` but without the
`Option`.

This drops out a few unnecessary `Option`-juggling in the deploy
code.

Prep for adding configmaps to the spec.

Signed-off-by: Colin Walters <walters@verbum.org>