tar: Fix multiple hardlinks
The previous commit handled the case of *one* hardlink,
but the logic for the secondary one was backwards.

I noticed this when generating a derived image, in some
cases we only end up with `/usr/lib/sysimage/rpm-ostree-base-db`
as modified, which causes really confusing semantics that
look like packages going missing.