install: Drop `--net=none` suggestion
While it's actually nice to show that the container can just
install itself, I'm trying to reduce the command line length
of the invocation in the interest of simplicity.

There's a minor cost here as it means we'll get a default
bridge network, but that hardly matters.  Anyone who wants to
avoid it can just specify `--net=none` on their own.

However this said, I think longer term we do want to support
networking at install time in order to do things like fetch
configmaps, so dropping this is prep for that too.

Signed-off-by: Colin Walters <walters@verbum.org>