install: Enable bootprefix by default
This is the last bit necessary to enable installations to
systems without a separate `/boot` partition (as is true in
the default RHEL 9 AMIs).

I was too chicken to enable it by default in ostree long ago;
and now here we are.  Should definitely enable it in ostree,
but let's do it here now.

Signed-off-by: Colin Walters <walters@verbum.org>