Improve EFI, add support for root=boot and writing UUID file
In FCOS we never tried to support root=boot, but for bootupd
to do alongside installs in the general case we have to.

There were two things to fix here:

 - Tweak the EFI "trampoline" to check for both $prefix/grub.cfg and $prefix/boot/grub.cfg
 - Add support for writing the boot UUID into both places

(This avoids higher level tools like bootupd needing to know
 about how to find the EFI vendor dir)

Then the bigger issue here is that we need to support invoking
`efibootmgr` to re-synchronize the firmware.  This is particularly
important in cases like bootc "alongside" installs where we're
taking over the target OS.