containerenv: small error message fix
This tweaks the error message when `.containerenv` cannot be found so
that a fully-qualified path is included in the error.

Previously, the error message would just show `run/.containerenv` as
the path to the file, which is not entirely clear.

I tried to change the `const PATH` to be an absolute path, but doing
so breaks the ability for `open_optional()` to open the file
successfully; returns an error `a path led outside of the filesystem`.

Signed-off-by: Micah Abbott <miabbott@redhat.com>