install: Just check if we're pid1
I'm working on a buildsystem where we run a "supermin" VM that
doesn't use systemd as pid1.  There's no reason to hard
require systemd for this.  This check isn't fully correct because
it will allow things to pass if someone does `podman run --init`
to make us pid2, but still in a userns.  In theory we could detect
that in the future.

Signed-off-by: Colin Walters <walters@verbum.org>