install: Re-implement hostns to work around supermin
I'm experimenting with doing `bootc install` in a supermin VM
like we do in coreos-assembler, but hit upon the fact that supermin's
`/sbin/init` implementation incorrectly does `chroot()` instead of
a `pivot_root`:

https://github.com/libguestfs/supermin/blob/5230e2c3cd07e82bd6431e871e239f7056bf25ad/init/init.c#L288

The problem here then is doing `nsenter` just takes us into the
original rootfs, and all the real content is in `/root`.

Re-implement this by re-executing our own binary and doing the `setns()`
and `chroot()` dance internally, because we can't really do it
with external binaries.

Signed-off-by: Colin Walters <walters@verbum.org>