Add convenient API/CLI to prune all content
This is kind of a usability regression from
https://github.com/ostreedev/ostree-rs-ext/pull/498/commits/979b93a24765ab129c6f1414388a9c7565fe3d80
"deploy: Add an API to prune undeployed images"

This new API and CLI make it convenient to prune the images,
the layers, *and* the underlying objects.

cc https://github.com/coreos/rpm-ostree/pull/4720

This API in particular is what bootc wants.  For rpm-ostree
we do still want to separate pruning refs from objects, because
rpm-ostree may have layered packages, and we only want to traverse
the ostree repo to prune objects once.