grubconfigs: move the GRUB2DIR creation earlier
We add the GRUB2DIR creation in ece9120 but there is an earlier
use of that directory if there are dropins so let's move the
creation of the directory earlier before the dropins are processed.

I hit this because I'm testing out things on FCOS. Building a qemu
and metal image works just fine, because in that case we are doing
the MBR install using `grub2-install` and I guess that tool creates
the directory. In the metal4k case, though, there is no BIOS install
and thus we get here and the GRUB2DIR doesn't exist and bootupd fails.