docs: Generate docs/man
This will help motivate us to improve our man pages, by making
them much more visible as pre-generated markdown that can
be conveniently viewed from e.g. the Github web UI.

For now, anytime we change the CLI we'll need to manually
run `cargo xtask man2markdown`, and commit and push the result.
This requires `pandoc` in the environment.

If this becomes burdensome I will likely investigate doing
e.g. a Github action which either automates doing it post-commit,
or enforces that it's done as part of the commit which introduces
the change.  However the problem with the latter will be the
`pandoc` requirement.

It may make sense to investigate e.g.
[clap-markdown](https://docs.rs/clap-markdown/latest/clap_markdown/)
in the future instead.

Signed-off-by: Colin Walters <walters@verbum.org>