Add `HostType`, change `status` to always succeed
Closes: https://github.com/containers/bootc/discussions/242

Basically it'd be useful for folks to be able to run `bootc status --json`
and have that always work - if we're not on a bootc-booted system
then we just output null data.

Specifically we drop the `isContainer` field and replace it
with an optional non-exhaustive enumeration.  For anything we don't
know about we just output `None` i.e. `null` in JSON.

Signed-off-by: Colin Walters <walters@verbum.org>