spec: Add cached update metadata
In https://github.com/ostreedev/ostree-rs-ext/pull/537 we
added a facility to cache metadata (manifest+config) for
container image updates with an eye for use in rpm-ostree.

The `bootc update --check` code here in bootc was also
updated to use it; but we didn't really expose it back in
the status.

This closes the gap, just bridging the cached update metadata
into status.  We want to do this as opposed to having
`bootc update --check` grow its own API for example.

Closes: https://github.com/containers/bootc/issues/247

Signed-off-by: Colin Walters <walters@verbum.org>