README: provide an example to generate a booting image
To help people get an impression of `bootc` this commit adds an
example how to install bootc to a disk-image and boot it with
qemu.

Arguable this example is not complete as there is no way for
the user to login in right now and maybe some example how to
use the host would be nice but this might be a useful start.