Add suggestion for `LABEL containers.bootc 1`
As we aim to slowly move away from ostree, let's add a bootc label
that we expect to be used.  For now, this is just a warning.

I will change rpm-ostree to add this by default if we take this.

But this one is also an important step for if we ever add the
flow of simply deriving from an existing application base image.

Signed-off-by: Colin Walters <walters@verbum.org>