store: If ostree >= 2024.3, retain content in `/var`
This is intended to pair with https://github.com/ostreedev/ostree/pull/3166

If we detect a new enough ostree version, then by default don't remap
content in `/var`, assuming that ostree itself will handle it.

In order to unit test this (without depending on the ostree version
that happens to be on the host) add an API to the importer
which allows overriding the version.