docs: clarify that `--karg` can be passed multiple times
It is not clear from the current man-page that `--karg` can be
passed multiple times. From reading the code it seems to be the
case but because it's not obvious (to me) I also added a small
testcase to the CI to ensure that I'm not misreading things.

I added a small comment to the docs that it can be given
multiple times.

Signed-off-by: Michael Vogt <mvogt@redhat.com>
Signed-off-by: Colin Walters <walters@verbum.org>