install: Add some more debug/trace calls
John is hitting an issue where we're failing to re-exec into
the mount namespace when querying podman for some reason.  We
log surprisingly little even at trace level before we get there.
I don't think this will help debug it, but I had to carefully
read the code to know that e.g. we've verified we're not running
under rootless podman until the time we get there.

More logging especially at trace level is ~free so let's
do it.

Signed-off-by: Colin Walters <walters@verbum.org>