install: Add `--skip-finalize`
A generic use case for installing to an existing filesystem
in the general case will be:

- `bootc install to-filesystem`
- Inject other configuration or data into `/etc` and/or `/var`

(Until such time as we offer an explicit API to do the latter)

Add a `--skip-finalize` which will avoid remounting the root.
This also should work around qemu-user not implementing the
FITRIM ioctl, which can be hit when doing cross-architecture
installs.

Signed-off-by: Colin Walters <walters@verbum.org>