bugfix: Switch losetup to --direct-io=off by default
When using a loopback device pointing to a file on a 9p filesystem (e.g.
the home mount of a podman machine), using direct-io=on causes the
partitioning via sgdisk to fail. This is a temporary fix until podman
machine switches away from 9p or another fix is found. Direct IO can be
enabled via the BOOTC_DIRECT_IO=on environment variable.

Fixes #485

Signed-off-by: Chris Kyrouac <ckyrouac@redhat.com>